sa-logic/sa/requirements.txt
Werkzeug==2.0
Flask==2.0.3
textblob==0.15.0
Pillow==8.4.0
requests==2.27.1
